Estudiantes y docentes

+++ config/filesystems.php
@@ -65,7 +65,8 @@ return [
             'host' => env('FTP_HOST'),
             'username' => env('FTP_USERNAME'),
             'password' => env('FTP_PASSWORD'),
-            'port' => env('FTP_PORT', 21),
+            // Ensure port is an integer because env() returns strings from .env
+            'port' => (int) env('FTP_PORT', 21),
             'root' => env('FTP_ROOT', ''),
             'passive' => true,
             'ssl' => false,
